+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [18.1.0](https://github.com/pgerke/ngx-mat-table-multi-sort/releases/tag/18.1.0) - 22.01.2025
+
+### Added
+
+- [#37](https://github.com/pgerke/ngx-mat-table-multi-sort/issues/37): Fix column reordering and visibility control styling
+- [#30](https://github.com/pgerke/ngx-mat-table-multi-sort/issues/30): Allow column reordering and visibility control
+- [#31](https://github.com/pgerke/ngx-mat-table-multi-sort/issues/31): Add an option to clear all sort criteria
+- [#28](https://github.com/pgerke/ngx-mat-table-multi-sort/issues/28): Add multi-sort control component
+
 ## [18.0.0](https://github.com/pgerke/ngx-mat-table-multi-sort/releases/tag/18.0.0) - 03.01.2025
 
 ### Added